requirements.txt
# Core dependencies
loguru>=0.6.0
PyYAML>=6.0
openai>=1.0.0
pydantic>=2.0.0

# HTTP and caching
httpx>=0.24.0
hishel>=0.0.24

# Web framework
fastapi>=0.100.0
uvicorn>=0.20.0

# ML and scientific computing
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
scipy>=1.10.0
scikit-learn>=1.4.0
datasets>=2.12.0

# Math and symbolic computing
sympy>=1.12
pylatexenc>=2.10

# Tokenization
tiktoken>=0.4.0

# Async and retry
tenacity>=8.2.0

# Progress bars
tqdm>=4.65.0
